+++ luscii_patient_actions_sdk_android/lib/luscii_patient_actions_sdk_android.dart
@@ -43,6 +43,11 @@ class LusciiPatientActionsSdkAndroid extends LusciiPatientActionsSdkPlatform {
   }
 
   @override
+  Future<void> logout() async {
+    await methodChannel.invokeMethod<void>('logout');
+  }
+
+  @override
   Future<List<dynamic>> getTodayActions() async {
     final actions = await methodChannel.invokeMethod<List<dynamic>>(
       'getTodayActions',