+++ luscii_patient_actions_sdk_platform_interface/CHANGELOG.md
@@ -7,10 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.11.0+1] - 2026-07-08
+## [0.11.0+1] - 2026-07-14
 
 ### Changed
 - Bumped version to `0.11.0+1` to align with the federated package release.
+
+## [0.10.3+1] - 2026-07-14
+
+### Added
+- Added the `logout` method to the platform interface.
 
 ## [0.10.2+1] - 2026-07-06